requirements/requirements.txt
django>=1.5

psycopg2
south

django-pagination
django_compressor
django-tastypie

python-memcached

PIL
sorl-thumbnail
mimeparse

django-twitter-bootstrap


#fabric